+++ CHANGELOG.md
@@ -17,6 +17,8 @@ Fixed Pylint bug with PyTorch 1.6
 Fixed `pack_padded` cpu error with PyTorch new release
 
 ## dev
+Added a verbose flag to trigger more message during embedding loading
+Added a verbose flag in model/embeddings download
 Added finetuning API to allow finetuning of our models (retrain)
 Added a method to test models (test)
 Added metric, loss and dataset container to facilitate finetuning training