+++ CHANGELOG.md
@@ -19,7 +19,7 @@
 - Fixed Pylint bug with PyTorch 1.6 
 - Fixed `pack_padded` cpu error with PyTorch new release
 
-## dev
+## 0.1.3.1
 
 - Added a verbose flag to trigger more message during embedding loading
 - Added a verbose flag in model/embeddings download